chore: Move IP utils to their own files
