+++ ui/src/components/EditDataNetworkModal.tsx
@@ -14,7 +14,7 @@ import { ValidationError } from "yup";
 import { updateDataNetwork, APIDataNetwork } from "@/queries/data_networks";
 import { useNavigate } from "react-router-dom";
 import { useAuth } from "@/contexts/AuthContext";
-import { ipv4Regex, ipv6Regex, cidrRegex, isValidIpv6Cidr } from "@/utils/bgp";
+import { ipv4Regex, ipv6Regex, cidrRegex, isValidIpv6Cidr } from "@/utils/ip";
 
 interface EditDataNetworkModalProps {
   open: boolean;
